requirements.txt
python-telegram-bot==20.6
mysql-connector-python==8.1.0
python-dotenv==1.0.0
PyMySQL==1.1.0
sqlalchemy==2.0.23
alembic==1.12.1
requests==2.31.0 